Refactor imports to use `config::AppConfig` instead of `AppConfig`
